Compile errors and warning fixes.
git-svn-id: https://zxtune.googlecode.com/svn/trunk@1235 e19a5227-2556-0410-816f-a35683127991

+++ src/core/plugins/players/ay/ay_base.cpp
@@ -305,7 +305,7 @@ namespace ZXTune
 
       void TrackBuilder::SetNoise(uint_t level)
       {
-        Chunk.Data[Devices::AYM::DataChunk::REG_TONEN] = level & 31;
+        Chunk.Data[Devices::AYM::DataChunk::REG_TONEN] = static_cast<uint8_t>(level & 31);
         Chunk.Mask |= 1 << Devices::AYM::DataChunk::REG_TONEN;
       }
